Refactor: Move home page and layout
