+++ frontend/src/routes/(app)/(home)/+layout.svelte
