+++ frontend/src/routes/(app)/(u)/(list)/accounts/+page.svelte
