--- frontend/src/tests/routes/app/root/page.spec.ts
@@ -1,10 +1,0 @@
-import App from "$routes/(app)/(u)/(accounts)/+page.svelte";
-import { render } from "@testing-library/svelte";
-
-describe("Layout", () => {
-  it("should render accounts landing page", async () => {
-    const { queryByTestId } = render(App);
-
-    expect(queryByTestId("accounts-landing-page")).toBeInTheDocument();
-  });
-});
